Working on Graph API, will continue when back home
