+++ .gitignore
@@ -52,3 +52,4 @@ android/app/libs
 fastlane/report.xml
 fastlane/Preview.html
 fastlane/screenshots
+.deco
